requirements.txt
# Core dependencies
pymongo==4.16.0
minio==7.2.20
kafka-python==2.3.0

# Web scraping
beautifulsoup4==4.14.3
requests==2.32.5
lxml==6.0.2
selenium==4.27.1  # For JavaScript-rendered content
webdriver-manager==4.0.2  # Automatic ChromeDriver version management

# Mock API server
flask==3.0.0

# Data processing
pandas==3.0.0
numpy==2.4.2
pydantic==2.12.5

# Fuzzy matching
rapidfuzz==3.14.3

# Workflow orchestration
snakemake==9.16.3

# Utilities
python-dotenv==1.2.1
pyyaml==6.0.3

# Visualization (optional)
matplotlib==3.10.8
seaborn==0.13.2
plotly==6.5.2
